Remove solver command
