+++ package.yaml
@@ -209,7 +209,6 @@ library:
   - Stack.Options.ResolverParser
   - Stack.Options.ScriptParser
   - Stack.Options.SDistParser
-  - Stack.Options.SolverParser
   - Stack.Options.TestParser
   - Stack.Options.Utils
   - Stack.Package
@@ -225,12 +224,9 @@ library:
   - Stack.Sig
   - Stack.Sig.GPG
   - Stack.Sig.Sign
-  - Stack.Snapshot
-  - Stack.Solver
   - Stack.SourceMap
   - Stack.StoreTH
   - Stack.Types.Build
-  - Stack.Types.BuildPlan
   - Stack.Types.CompilerBuild
   - Stack.Types.Compiler
   - Stack.Types.Config
