--- test/integration/tests/3607-stack-init-solver/files/.gitignore
@@ -1,4 +1,0 @@
-.stack-work/
-foo.cabal
-*~
-stack.yaml
