--- test/integration/tests/3607-stack-init-solver/files/package.yaml
@@ -1,8 +1,0 @@
-name:                foo
-version:             0.1.0.0
-
-dependencies:
-- base
-- acme-missiles
-
-library: {}
