--- test/integration/tests/cabal-solver/files/.gitignore
@@ -1,2 +1,0 @@
-stack.yaml
-.cabal-sandbox
