--- test/integration/tests/cabal-solver/files/Foo.hs
@@ -1,1 +1,0 @@
-module Foo where
