--- test/integration/tests/cabal-solver/files/foo.cabal
@@ -1,10 +1,0 @@
-name:                foo
-version:             0.0.0
-synopsis:            foo
-build-type:          Simple
-cabal-version:       >=1.10
-
-library
-  exposed-modules:   Foo
-  build-depends:     base, acme-dont == 1.1
-  default-language:  Haskell2010
